+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Make `fetch_object` return `JsonValue`.
+
 ## [0.3.0] - 2024-10-16
 
 ### Added