+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-11-07
+
 ### Added
 
 - Added `skip_verification` parameter to `fetch_object` options.